[editorconfig] make npm package published as public
For Theia extensions, there is an npm setting in package.json, to select
whether a package is published on npm with public or restricted access.
By default our scoped packages are be published with restricted access.
This fails unless a project has a paid npm account. So we want to publish
our extensions as public. It seems this was missed, in this new extension,
preventing it from publishing successfully to npm, in the post-merge builds.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>