[core/cluster] Try to close each process gracefully
When the backend closes, processes spawned when running in cluster mode don't always close properly, making the terminal where Theia was started hang.
Cause can be unfulfilled promises keeping the event-loop alive.

Solution was to make sure that in case an error happens, it is detected and call `process.exit()`.

Fixes #1883

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>