GH-1924: Fire a selection change event when calling `addSelection`.
Even if the selection does not change under the hood. This is to ensure that
the selection is correctly propagated to the selection service at
application startup. With this change we, get the correct context menu
contribution in the navigator.

This changes also sets the focus on the tree widget before restoring the
selections. So when the layout is reset, and the user activates the `Files`
widget, the selection is available for the context menus.

Closes #1924.

Signed-off-by: Akos Kitta <kittaakos@gmail.com>