Fix stubbing in keybinding.spec.ts
In patch "keybinding.spec.ts: Remove unnecessary stub variable", I
removed a line that stubbed os.isOSX to false, because I thought it was
done in the beforeEach.  However, the file is quite long and I did not
notice that the test from which I removed the stubbing was in a
different section ("describe") than the beforeEach that did the
stubbing.  It was therefore wrong to remove it, so this patch reverts
that.

However, to make things clearer, we should declare and use stub
variables in the smallest scope possible.  This will help avoid errors
like that, where the same variable was used in two completely different
contexts.  So this patch declares different stub variables in narrower
scopes.

Finally, I think it is wrong to do:

  const obj = {
    prop: 1
  };

  stub = sinon.stub(obj, 'prop').value(2)
  stub = sinon.stub(obj, 'prop').value(3)

  stub.restore();

Because the value restored will be 2, not 1 (the original one).  In this
case, we should remove the last assignment to stub, so that we will
correctly restore the original value.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>