GH-2042: Added a default `Mini Browser` endpoint handler.
So even if a given file cannot be opened in the `Mini Browser`
it can be still handled by sending back its content to the frontend.

This fixes the issue when trying to load the CSS files for a local HTML.

This commit also fixes the `.gif` support and uses a proxy service (over
JSON-RPC) to populate the supported file extensions, insted of calling
a static endpoint from a browser `fetch`.

Closes #2042.

Signed-off-by: Akos Kitta <kittaakos@gmail.com>