[terminal] Fix #2045: simplify and fix resizing logic
The main issue is a race between resizing a terminal and opening a new terminal process. Calling resize on terminal does nothing if a terminal was already resized with such cols and rows, because of it a newly created process does not get proper sizes even that the terminal resizing is called.

Signed-off-by: Anton Kosiakov <anton.kosyakov@typefox.io>