Adjust our tslint config to tslint defaultSeverity behavior change
We have two set of lint rules that we want as either warnings or errors
in configs/warnings.tslint.json and configs/errors.tslint.json.  Each
file specifies the "defaultSeverity".  With tslint < 5.9.0, the
defaultSeverity would be applied to rules in that file, so everything
worked happily, we had warnings and errors reported in the IDE.

Starting with 5.9.0 [1], tslint changed how it handles defaultSeverity
in extended-from config files.  Now, the defaultSeverity of the last
included file wins.  So with our config and tslint 5.9.0, rules
specified in errors.tslint.json appear as warnings, since
warnings.tslint.json is included last in the top-level tslint.json.

To fix this, removed the defaultSeverity from warnings.tslint.json and
specified it explicitly for each rule.  It's more verbose, but the
number of rules we want as warning should decrease with time (when we
fix all the violations of a rule, we move it to errors.tslint.json).

I verified this by introducing an error (removed the copyright comment
from a file) and running tslint before and after.  This is the
difference:

-WARNING: (file-header) /home/emaisin/src/theia/packages/callhierarchy/src/browser/callhierarchy-service-impl.ts[1, 1]: missing file header
+ERROR: (file-header) /home/emaisin/src/theia/packages/callhierarchy/src/browser/callhierarchy-service-impl.ts[1, 1]: missing file header

The "curly" rule is present in both warnings and errors, so I removed it
from warnings.

[1] https://github.com/palantir/tslint/releases/tag/5.9.0

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>