tslint: Enable trailing-comma
I like when people use a trailing comma in comma-separated definitions.
Turns out there is a tslint rule for it!  And it also has an autofixer,
so I ran tslint --fix to generate this patch.  I sampled a few files and
the results seems good.

For example, I would prefer this:

const foo = [
  A,
  B,
  C,
];

rather than:

const foo = [
  A,
  B,
  C
];

The main reason is that it simplifies diffs where you add a new item.
You don't need to change the line containing "C".

With the "multiline" setting, things like this are not affected:

const foo = [ A, B, C ];

If you want to add "D", you'll need to change the line anyway, so the
comma would not help here.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>