cpp: Send clangd compilationDatabasePath in 'initialize' request
That way, as soon as ths initialize is received by the server, it can start
indexing with a valid compilation database and not have to wait for a
'didChangeConfiguration' that might or might not happen.

I was tempted to keep the initial "didChangeConfiguration" so that it doesn't
break existing Clangd but there was no released versions yet with this support
anyway. So, provided that we can change Clangd before next version (branching)
then we don't have to keep this compatibility bit.

Depends on https://reviews.llvm.org/D49833

Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>