[ui tests] remove PhantomJS dependency
The PhantomJS npm package is deprecated, since the development has
stopped a few months ago. We have a dev-dependency for
wdio-phantomjs-service, in the root package.json. In turn, this
pulls PhantomJS. However it seems that we are not actually using it.
This patch removes this dependency.

see: https://github.com/ariya/phantomjs/issues/15344

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>