Fix workspace uri long names when path is not really in workspace
If the workspace root is /tmp/foo and the file to get the long name of
is /tmp/foo-bar/baz, we currently wrongfully shorten to -bar/baz.
Change the code so it matches a directory separator after the workspace
path.

This patch also adds a test.  In order not to have to mock too much
things, I extracted the necessary bits of WorkspaceService (that are
used in WorkspaceUriLabelProviderContribution) in a new
IWorkspaceService interface.

Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>