Run UI tests headless by default
I often run the testsuite and do something else while it runs in the
background.  Having the browser windows pop up randomly is annoying, and
it actually risks interfering with the tests, if I accidentally click or
type in the window.  Therefore I suggest making them headless by
default.

A new "yarn run test-non-headless" target is added to the package.json,
in case somebody wants to run with the browser windows visible.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>