task-server test: Remove setting of process.argv
I believe that this modification of process.argv is not useful.  I
tried to change the --root-dir value to a random directory, and the test
still passes.

What hinted me is that the code that consumes process.argv in Theia is
the CliManager class, but it is not even instantiated during these
tests.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>