Make some waitFor* functions use waitForVisible
I am sometimes seeing some random failures in the UI tests.  I believe
that this will help (or even if it doesn't help, I think it is right).

When a button is present on the left or right tab bar, but the
corresponding view is closed, the node for that view is still present in
the DOM.  For example, even when the Files view is collapsed, the #files
div is present in the DOM, it's just not visible.  Therefore, the check
waitForExist('#files') in waitForFilesView isn't right, as it won't wait
until the Files view is open (which is what I assume is the goal of that
function).

This patch replaces waitForExist with waitForVisible, to make sure we
wait for it to be visible.

I also changed the names to be clearer about what the functions do, for
example waitForFilesView becomes waitForFilesViewVisible.

Change-Id: I3ba1d3899fe161c7c14ce1de47dd450d2b0842e0
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>