Run examples/browser UI tests using a temporary workspace
I am looking to add a UI test for some C/C++ features.  That will
require writing some files to the workspace and open them in the editor.
Right now, the workspace used for these tests seems to be based on the
content of ~/.theia/recentworkspace.json.  We obviously don't want to
start writing stuff in the user's real workspaces, and since we don't
know what we will find in there, it could make the tests unstable.

This patch makes the UI tests run with a temporary workspace, deleted
after the test run.

All the UI tests still run simultaneously using the same
workspace/backend, so there is still room for unpredictability.

Change-Id: I9c140daa774cc5f67f80fc10302f7b34efc057a6
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>