Make it possible to pass process args when starting a backend
When doing some tests that require starting a backend (in particular the
UI tests, but not only them), we sometimes need to pass arguments to
that backend.  Right now, the only way is to modify process.argv, but
it's fragile because we don't know yet what is there in the first place
(it's the arguments to the test process).

This patch makes it possible to pass args to the backend start
function.  In the normal case, we pass process.argv, but in the tests,
we can pass an arbitrary array.

Change-Id: I2ed2308509dd476ab4d1506482075863fc4e84eb
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>