Restart clangd when changing build configuration
Currently, when the user changes build configuration, we send a
DidChangeConfiguration notification to clangd.  This does not work in
the particular case where the user wants to go back to the default of
not using a specific build configuration.  There is no way (as of now)
to tell clangd that we don't want to use a specific
compile_commands.json anymore.

This patch works around it by restarting clangd when the user switches
config.  The new clangd will be initialized with the newly selected
build config.

The only way to test this end to end (with clangd involved) was to add a
UI test.  The test is skipped if clangd is not available, since we don't
expect everybody to have it installed.  If/when the cpp extension is
moved to its own repo, then we can make the test unconditional.

Change-Id: Id4e42ccd3b69a147a4c099e73440b8df00981e67
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>