Remove 'Current Workspace' label from Open Recent Workspace
Since the file `recentWorkspaces.json` is modified each time an instance of Theia closes, it
is not always the case that the first element is the most recent workspace but rather the last closed.

Removing the label `Current Workspace` and instead always displaying the last modified date
removes any ambiguity.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>