Improve output widget readability
Because the output view uses renders text as HTML inside a simple div,
any leading whitespace is ignored and any repeated whitespace appears as
just one.  This makes formatted output from tools very difficult to
read.

This patch makes it so whitespaces are rendered by the browser (using
the "white-space: pre" css property.

It also makes the font monospace, since the text we find in the output
widget is most likely to be more readable when using a fixed-width font
(the language server output, for example).

Change-Id: I62b30d5a27222cb0cf851962c5a7aaea2c533ad5
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>