Add no-implicit-dependencies tslint rule
This rule requires that if a package imports something from another
package, the latter must be present in the former's dependency list, in
the package.json.

Currently, it's possible to use something in package A from a package C
that was pulled in indirectly by another package B. When trying to make
an application with A but not B, it's possible that C will not get
pulled in (because nothing in that application explicitly requires it)
and we get a build error.

Another typical mistake that this rule catches (a bit by chance ) is to
use "import ... from '@theia/foo'" from source code in the @theia/foo
extension. In my experience, this may cause some strange build failures,
so it's better to use relative imports in that case.

I tried to make a reasonable whitelist with the packages that we don't
want to re-declare in every package.json.  For example, inversify is at
the base of every Theia application, so I can't imagine a build of Theia
that will not include it.  Test tools such as chai are declared in the
root package.json, so we can assume they will always be there.

In other cases, it's better not to declare the same dependency twice,
because it would cause complicated version conflicts.  One example of
this is the vscode-languageserver-types package.  If we specify the
dependency multiple times using different versions, we will get some bad
surprises when different packages try to pass around object types from
different versions of the library.

Since this only generates warnings, we can take the time to fix the
issues one a the time.  I think all offenders should fall in one of
these categories:

1. The dependency is indeed missing, so we add it.
2. The dependency should by whitelisted, we add it to the whitelist.
3. This particular instance should be ignored, but should not be
   whitelisted at large, so we add a "tslint-disable" comment.

Change-Id: I246937c424ed56a57ffda19567f3927fcdccb57b
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>