Fix task exit messages and task/process API
The original intention for this patch is this fix the message that
appears when tasks exit.  There are many inconsistencies:

task of type shell exits with exit code 0:
  Success - Task 2 has finished. exit code: 0, signal: undefined

task of type process exits with exit code 0:
  Success - Task 7 has finished. exit code: 0, signal: null

task of type shell exits with exit code 2:
  Error: Task 1 failed. Exit code: 2, signal: undefined

tas of type process exits with exit code 2:
  Error: Task 8 failed. Exit code: 2, signal: null

task of type shell terminated by signal SIGUSR2:
  Interrupt received - Task 3 has finished. exit code: 0, signal: 12

task of type process terminated by signal SIGUSR2:
  Error: Task 5 failed. Exit code: null, signal: SIGUSR2

There are a some problems:

1. There is always either an exit code, either a signal, never both.  So
   we don't need to show both (the "undefined" or "null" are particularly
   annoying).
2. We should show signal names for all signals.  That signal name should
   be provided by the backend and not interpreted by the frontend,
   because they vary per OS.
3. Whether receiving a signal is an error is debatable, some signals can
   be expected.  Receiving a SIGSEGV is probably not, though.

There are a lot of inconsistencies with how node's child_process module
and the node-pty module report their exit events.  This patch tries to
normalize them so we get something predictable in the frontend.

End-to-end, I made the "code" value optional, so that when a signal is
reported, code will be undefined (compared to now where it will be 0).

Random notes:

- I found setting "strictFunctionTypes" in the compile.tsconfig.json was
  helpful for this, and it can only help in general, so I suggest we
  leave it.
- Object.entries is only available in ES2017, so I added a small
  function that does the equivalent.
- I removed the messages like "Interrupt received" in the messages.  I
  think they were a bit misleading.  If we want to have an "official"
  signal description, we should use the descriptions returned by
  strsignal(3).

Change-Id: I20d2dec8e131d95018cc6f1bf7bdf1876c662811
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>