Disable tslint indent rult for json-schema.ts
This file uses tabs instead of 4 spaces, which our tslint.json file
specifies.  Since this file was copied from the vscode repo, I chose to
leave the tabs there and just disable the rule for this file.  This will
make it easier to diff it with the original file, in case we need to
update it.

The warnings are not shown by tslint (probably because of
https://github.com/palantir/tslint/issues/2814), but they show up in our
Sonarqube project.

Change-Id: Idc7536e5e80be4d7e0d32b529a442d99365f7a19
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>