Improve prepare-travis for Windows
When running "yarn" on Windows, we get spurious diffs in .travis.yml.
The first reason is that since we use path.join, it replaces all / with
\.  Fix this by joining with '/' all the time.

Then, we get some git warnings about LF/CRLF.  The reason is that with
git on Windows, it's common to use core.autocrlf=true.  This means that
files are saved with \n end of lines in git objects, but those are
replaced with \r\n when checked out on Windows.  When we re-generate
.travis.yml on Windows, we currently, replace the \r\n (in the checked
out version of the file) with single \n.  Then, git warns us that our
file changed, but not really:

    $ git status .travis.yml
    ...
            modified:   .travis.yml
    ...
    $ git diff .travis.yml
    warning: LF will be replaced by CRLF in .travis.yml.
    The file will have its original line endings in your working directory.

Fix this by using platform-specific end of lines.  With this, running
"node scripts/prepare-travis" on a clean Theia repository results in no
changes to .travis.yml, on Linux or Windows.

Fixes #3171

Change-Id: I15e6e1ffed6be4aa3830c1c59167f7f9e91d50a3
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>