Fix races in kill tests in task-server.slow-spec.ts
The two tests modified in this patch are racy (I get some random
failures).  Depending on the timing, we can kill the process before we
have installed the "onTaskExit" handler.  This results in the handler
never being called, and the test fails after the 10s timeout.

This was found by working on Windows, but can easily be replicated
anywhere by calling taskWatcher.onTaskExit in a setTimeout(..., 2000) to
delay it on purpose.

To fix it, send the kill command only after we have installed the
handler.

Also, if the exit event we receive doesn't match what we expect, call
reject().  This makes the test fail quickly, rather than wait for the
timeout.

Change-Id: I7a91ec6744fda562cdb3676666e29b557f5dcd35
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>