dugite-git.spec.ts: Increase timeout for 'log' suite
Some git tests often fail on AppVeyor.  For example:

    @theia/git:   1) log should not fail with local git when executed from the repository root:
    @theia/git:      Error: Timeout of 2000ms exceeded. For async tests and hooks, ensure "done()" is called; if returning a Promise, ensure it resolves.

The 'git' suite, higher in the same file, already has a
this.timeout(10000).  I think that the test that fails here would pass
if we gave it a bit more time (since it works intermittently).  So this
patch sets the timeout to 10000ms for the 'log' suite as well.

Change-Id: I0f676313f12066c69df5cd3bef1d0bcd18e1d004
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>