Make process creation asynchronous, improve error reporting
The error reporting when spawning a process fails (notably ENOENT and
EACCES) is not very good currently.

For example, after trying to execute a non-executable file, we get a
RawProcess object with the `process` property undefined.  When we try to
execute a non-existent file, we get a RawProcess.process with an
undefined pid.  I think it's not very practical to be able to receive
RawProcess objects that represent processes that we failed to start.

With this patch, I propose to extract the actual process spawn from the
RawProcess and TerminalProcess constructors and move that in some
factory methods that return Promises.  The RawProcess and
TerminalProcess objects will be created to wrap successfully started
processes.  Errors will be propagated by returning rejected Promises.

Doing this change has quite a bit of impact on other modules, but I
think that in the end it's an API that's harder to misuse.

Change-Id: I76bc0cc5ab0664c04ddc73e4d68e78948780f9d0
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>