GH-3397: Implemented the HTTP-based authentication for Git in Electron.
If a Git operation (fetch, pull, merge, ...) requires authentication,
the Theia backend will ask the frontend for the username and password.

User credentials are neither stored nor reused.

In electron, if a Git operation fails due to authentication, we
report it back to the user. This PR does not change the behavior of the
browser-based application.

Closes: #3397

Signed-off-by: Akos Kitta <kittaakos@typefox.io>