Add error case tests for search-in-workspace test
Add tests for when the rg binary is not found and is not executable, to
make sure we handle that correctly.

Note that when running the test case, I get a backtrace:

  ripgrep-search-in-workspace-server
TypeError: Cannot read property 'kill' of undefined
    at RawProcess.kill (/home/emaisin/src/theia/packages/process/src/node/raw-process.ts:103:26)
    at ProcessManager.unregister (/home/emaisin/src/theia/packages/process/src/node/process-manager.ts:62:21)
    at /home/emaisin/src/theia/packages/process/src/node/process-manager.ts:46:36
    at CallbackList.invoke (/home/emaisin/src/theia/packages/core/src/common/event.ts:107:39)
    at Emitter.fire (/home/emaisin/src/theia/packages/core/src/common/event.ts:199:36)
    at /home/emaisin/src/theia/packages/process/src/node/raw-process.ts:92:35
    at _combinedTickCallback (internal/process/next_tick.js:132:7)
    at process._tickCallback (internal/process/next_tick.js:181:9)
    ✓ fails gracefully when rg isn't executable

This is due to a latent bug in RawProcess, where if we try to execute a
non-executable file, the process property of the RawProcess is
undefined.  It doesn't make the test fail, however, and exists prior to
this patch, so I decided not to address it right now.

Change-Id: I19d95d9b003a48096c14a51dd81d8b9c48fdde59
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>