[electron] Remove default menus on startup
When we open an Electron window, a bunch of menus are defined by default.  This
is annoying because when we want to quickly open a workspace, we might click on
those, then once the application really starts it will destroy the menus and we
have to clic back on it to finally do what we wanted in the first place.  This
commit removes the menus until the application sets them back.  The generated
code will also look for an `applicationName` field inside the `package.json` of
the application as title of the created window, again until everything is
loaded.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>