[electron][navigator] Add `OpenFolder` button
Only the `Open Workspace` button is shown in the navigator.  But since native
file dialogs were added, opening a workspace specifically requires you to open
a file. This was a bit annoying.

Also because of #3216 we can corrupt the current workspace and be stuck with
it...

This commit only change the displayed button to trigger `Open Folder` instead
of `Open Workspace` under Electron + Unix/Windows.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>