Fix loggers config file not taken into account
The ConsoleLoggerServer class fails to honor the initial settings in the
file specified by --log-config.  In init(), it correctly saves the right
log levels for each logger.  However, when a child logger is created, it
overrides the log level (in its local loggers map) for that logger with
INFO.

ConsoleLoggerServer probably doesn't need to save the log levels, since
it just replicates what LogLevelCliContribution already does.
Therefore, remove the loggers map and use LogLevelCliContribution
directly when needing to query for a log level.

Change-Id: I14a017dede54fd64a67f8c51e32a076403b5d2cb
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>