Fix check-hoisting not reporting diagnostics
I tried to use check-hoisting and found that it didn't report anything,
in situations where it clearly should have (I had different versions of
@theia/core).  The line with `issues.length` is wrong, because issues
is a Map object.  It should be `issues.size`.

It also found a discrepancy between toSeverity and toType, where one
uses `'warn'` and the other `'warning'` (they should both use the same
value).

Since this is the kind of problems that would have been caught by
writing the script in Typescript, I converted it to Typescript (ran with
ts-node).

Change-Id: Id65ba78835eef5b369b0486643f9eeffa1bef58e
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>