Fix rule definitions in errors.tslint.json/warnings.tslint.json
Opening error.tslint.json/warnings.tslint.json in Theia, I noticed it
complains about them not matching the expected schema.  It turns out
that the error ones should not have the "options" field.  For the
warnings one, we do not need to specify "true" if we provide options for
the rule (it's implied).

This patch fixes that, as well as some files that were infringing the
indent rule (treated as an error).  I verified that we get the same
warnings as before.

Change-Id: Iec23facc5ca4cca4e22cf964f873ef37d59ffe37
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>