Adjust AlertMessage font-color for bright themes
Before the change, the color was determined through variables defined
through the dark and bright variables.css but this should not apply
for `AlertMessage`s since we control the background color and want
the proper contrast. If a white theme was used before, we displayed
black font which was not the intended or wanted styling.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>