[monaco] Add `onWillChangeModelOnSave` event
This commit addresses the issue when some contributions want to actually change
a document, without providing changes to the API and let monaco do.

If someone changes the document after another contribution already provided
edits, it might offset every previously provided edits.

By adding the new event, we can ensure that every change that is done in
one-shot will not interfere with the other ones, by executing them
sequentially, so that we smoothly get from some "Version A" of a document to
"Version B", then from "Version B" to "Version C", etc...

Once the changes are made, people using the API to provide edits to be made by
monaco can be ran concurrently.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>