[Keybinding] Normalize key sequences to US layout
Some bindings, i.e. 'Toggle Line Comment' or 'Open Terminal' are
bound to a sequence that are not directly reachable on many
non-US keyboard layouts. For instance 'ctrlcmd+/' to toggle line
comments doesn't work on german kb layouts, because the slash
is above the `7` and therefore the shortcut is seen as `cmdctrl+shift+7`

This change translates the given chord to
a normalized US keysequence if the character is one of the US layout
keys. I.e. for `cmdctrl+shift+7` we also know the typed character is `/`
so if there is no direct keybinding we test against a US-keyboard
layouted version.

Fixes theia-ide/theia#1244

Signed-off-by: Sven Efftinge <sven.efftinge@typefox.io>