Fix clangd being started early when a build config is active
If a C/C++ build config is active, clangd will be started when the
workspace is started, even if no C/C++ file is open.  The reason is that
we emit an onActiveConfigChange event on startup, to tell the world
about the initially selected config.  This calls "restart" on the
language client contribution, and causes clangd to be started.

Fix this be first checking if clangd is already running before calling
restart.

Fixes #3853

Change-Id: Ie2bef70b42da44ec096aee0d404aefd0f666c73a
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>