add multi root support to WorkspaceService.getWorkspaceRootUri()
WorkspaceService.getWorkspaceRootUri() returns the first root folder that the given file belongs to, which is correct but inconsistent with other services of theia (e.g., search-in-workspace) when the file can be found in more than one root folders. With this change, the possibility of having multiple root folders is taken into account.

Signed-off-by: elaihau <liang.huang@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>