[filesystem] `createFolder` to not fail if exists
When we want to create a folder, we usually do not care if the folder
already exists or not, if we did, we could always query it using
`filesystem.exists`. Currently, `filesystem.createFolder` will fail if
it already exists.

This commit makes it so `createFolder` will not fail if the file that
exists is also a folder.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>
Signed-off-by: Bogdan Stolojan <petre.stolojan@arm.com>