[filesystem] Run only one file watching process
`FileSystemWatcherServerClient` is not bound as a singleton, issue is
that for each instance of that class, a new process is created.

So the backend would usually have one file watching process opened, no
frontends.  Then, for each new frontend, 2 more processes are created
for file watching.

With this commit, everyone should use the same watcher process.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>