Implemented Toolbar support for sidepanels and changed sidepanel tabs.
Every sidepanel has an header now where the title and a toolbar
is shown (if tools are registered for that widget).
As an example the control buttons of search in workspace extension are
moved to the toolbar.

Additionally the tabs of sidebars  don't show labels anymore but icons.

Fixes theia-ide/theia#3864

Signed-off-by: Jan Bicker <jan.bicker@typefox.io>