[process/task] Normalize tasks types and processes
When executing tasks, we now always use node-pty.

The difference between a process or a shell task relies in the way the
command and its arguments are interpreted/executed:

- process: ask the system to spawn a specific executable, with args
- shell: defer the spawning to a shell that will evaluate a command line

When spawning a process, the expected paramaters are of the kind:

	spawn(executable, [arg1, arg2, arg3, ...])

While when spawning a shell task, we expect some command like

	$> executable arg1 "arg 2" ... && sleep 1 || echo fail

This commit makes Theia tasks and "terminal processes" (processes
spawned via node-pty) able to use both inputs.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>

[process] Mark as killed when process errors out

Removes some unhandled errors from the tests because we used to not mark
processes that errored as `killed`.

Also renamed a potential name collision with the global `process` name.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>

[process] Add support for arguments in shell task

VS Code supports an extra `args` field for shell tasks, and will try its
best to build a command-line as `task.command ...task.args`, it will
also take the care of escaping some arguments. It is also possible to
explicitly define what escaping style you want the task runner to use.

This commit makes a best effort to comply to most use cases.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>

---

IMPORTANT: Disabled `Terminal Process { shell: true }` on Windows.

Problem is that only in this scenario, a lot of garbage ends up in the
standard output of the processes, and I have no idea why. In practice
this doesn't impact the actual terminals inside Theia.

So the tests are disabled on Windows by default, you can have them run
by setting an environment variable named `THEIA_PROCESS_TEST_OVERRIDE`.