Set the minimum perfect scrollbar length
Fixes #2788

Set the minimum perfect scrollbar length with the `minScrollbarLength`
scrollbar option. Currently in Theia, when there is a lot of vertical
height the scrollbar becomes quite small and hard to use. With this
change we define a minimum length for the scrollbar so it always
remains easily visible for users.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>