Added NOTICE.md file
This file is mandated by the Foundation. For the most part it's auto-
generated. Exception: the "Electron" section towards the end has been
manually added, as requested in CQ #19253, comment #17.

To eventually update this file, See the Eclipse Theia project page,
navigate to the "Legal Documentation Generator" and look for "NOTICE
FILE".

Today the direct URL is:
https://www.eclipse.org/projects/tools/documentation.php?id=ecd.theia

Make sure to preserve the manually added "Electron" section
at the end.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>