Implement 'Save As'
Fixes #799

This PR consists of:

1. Implemented the `Save As` functionality and command in order
to save a source URI to a target destination. The `URI` in question
can either be a single file URI or a folder URI (save all files under the
directory to a target).

2. Updated the `SaveDialogProps` to be able to pass an `inputValue` to
the html text input. This is useful when providing a programmatic name
to the input if available. In the instance of a `Save` dialog it is used
to show which `URI` currently has the context.

3. Fix bug when rendering the `SaveDialog` tree icons.
Incorrectly we were using the passed `URI` to determine the correct root icon
instead of using the root `FileStat` which provides the appropriate icon.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>