Add ability to remove recent workspaces
Fixes #4884

- Added the ability to remove `Recent Workspaces` from the
`Open Recent Workspace...` quick-open command palette.
At the moment, the only way to be able to remove workspaces
from the list was to completely delete the `recentworkspace.json` file
found under the user's home under `.theia`. This meant that users
would not be able to trim their list of outdated workspaces causing
too much noise. Users can now directly remove items from the list.

- Fixed issue with the `file-icon`. Passing a `URI` to the label
provider to generate the file icon is troublesome as it looks simply
at the extension. In the case of a folder like `.theia`, the icon is incorrect.
Instead, the `FileStat` should be used which provides more information.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>