GH-4940: Changed the way we require the server.
Instead of resolving the path of the server module and requiring it in
the Theia's `main`, we require the server module directly.

This change does not intend to change the actual behavior but supposed
to help the packaging process when bundling Theia into a single
executable.

Bundlers usually parse `requires` from the source code and they cannot
do much out-of-the-box with requires called with a non-literal
argument.

Closes: #4940.
Signed-off-by: Akos Kitta <kittaakos@typefox.io>