Allow to skip npm `prepare` script
Sometimes we just want to update dependencies, and building Theia is
really not in our intent.

This commit adds a check on a `THEIA_SKIP_NPM_PREPARE` environment
variable to suppress the automatic building.

Note that it will keep executing some scripts, but they don't take a lot
of time, unlike a full project build.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>