Enable navigating markers with single-click
Fixes #5003

Enable navigating between markers in the `problems-widget`
with the use of a single-click rather than always a double-click.

Folder nodes behave the same way where a single-click collapses them,
while markers now simply require a single click (`select`) in order
to navigate to the corresponding marker in the editor.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>