Filter out preferences without group names
Fixes #5066

Currently, if a preference in Theia is malformed, the `preference-tree`
will not display correctly. Instead of adding an invalid preference group
name, filter the names out of the tree completely. In addition, preferences which statisfy
the condition `isSectionName` should not be added to the tree since their `Add Value`
functionality does not work.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>