[electron] Test codecs from ffmpeg
Electron's ffmpeg can bundle proprietary codecs, which we want to avoid
having.

This commit adds a native module that will be built and run on the
client where the `@theia/electron` package will be installed. The
program will link against Electron's ffmpeg shared library and scan the
available codecs, failing if a proprietary one is found.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>