Optimize build (removing comments + helpers)
When transpiling TypeScript sources into JS (ES5), the compiler add a
bunch of helper functions, such as `__extend`, `__await` or others. This
is required for our code to work with older Javascript targets.

Problem is that these helper functions are added to each and everything
file that requires it, and when bundling the frontend application it
doesn't get refactored either.

But TypeScript allows us to depend on a package that implements all the
helper functions in one place, and then replace each definitions in the
file headers by 1 line that requires the package.

This commit also removes the comments when transpiling, since it ins't
that useful at runtime (we still have source-maps)

Size difference of some folders:
- `packages`: -4MB
- `examples/browser/lib`: -2MB

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>