Remove separator when no recent results are found in the quick-file-open
Fixes #5547

- Currently, when no recent results are found when performing the
quick-file-open, we display a separator incorrectly. The separator should
only be displayed to differentiate between two groups of items and thus
should only be displayed if there currently are recent results. The PR
addresses the issue by only displaying the separator if the recent results
list has one or more values.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>