Add the ability to 'save and retry' when prompted for dirty preferences
Added a small functionality to the message displayed to users
when there are dirty preference editors and they attempt to make
an edit through the preferences-tree. Users will have the ability
to select the `Save and Retry` button to successfully save their
preference editors and re-attempt to apply a preference edit.

- added small fix where the editor was not updated immediately after
setting a new value from the preferences-tree

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>