Do not re-open an already currently opened workspace
Fixes #5630

- fixes an issue that allows users to open a workspace
that is currently opened. If the user had the preference
`"workspace.preserveWindow"` set to `false` it would open
an identical workspace in a new tab, and if it was set to
`true` it would reload the workspace. Instead, an additional
check is performed to determine if the current workspace and
the new destination are equal, and if they are a NOOP is
instead performed by returning `undefined`.
- fixes issue present in `OPEN` and `OPEN_WORKSPACE` for both
the browser and electron based applications.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>